requirements.txt
Pillow>=10.0.0
tkinter
numpy>=1.21.0
